Excluding more files.
