+++ src/Hyjinx.LibHac/Crypto/ICipherWithIv.Legacy.cs
@@ -1,0 +1,12 @@
+﻿#if IS_LEGACY_ENABLED
+
+using LibHac.Common;
+
+namespace LibHac.Crypto;
+
+internal interface ICipherWithIv : ICipher
+{
+    ref Buffer16 Iv { get; }
+}
+
+#endif
