+++ src/Hyjinx.LibHac/FsSrv/Storage/MmcService.cs
@@ -13,19 +13,6 @@ using MmcPartition = LibHac.Fs.MmcPartition;
 namespace LibHac.FsSrv.Storage;
 
 /// <summary>
-/// Contains global MMC-storage-related functions.
-/// </summary>
-/// <remarks>Based on nnSdk 16.2.0 (FS 16.0.0)</remarks>
-public static class MmcServiceGlobalMethods
-{
-    public static Result GetAndClearPatrolReadAllocateBufferCount(this FileSystemServer fsSrv, out long successCount,
-        out long failureCount)
-    {
-        return fsSrv.Storage.GetAndClearPatrolReadAllocateBufferCount(out successCount, out failureCount);
-    }
-}
-
-/// <summary>
 /// Contains functions for interacting with the MMC storage device.
 /// </summary>
 /// <remarks>Based on nnSdk 16.2.0 (FS 16.0.0)</remarks>
