+++ src/Hyjinx.LibHac/Tools/FsSystem/ConcatenationStorageBuilder.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿#if IS_LEGACY_ENABLED
+
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using LibHac.Fs;
@@ -63,3 +65,5 @@ public class ConcatenationStorageSegment
         Offset = offset;
     }
 }
+
+#endif
